Move integration test out to proper location.
