+++ src/lib.rs
@@ -1,0 +1,16 @@
+mod config;
+mod domain;
+mod error;
+mod server;
+mod server_tracing;
+mod util;
+
+pub use config::{
+    NotaryServerProperties, NotarySignatureProperties, ServerProperties, TLSSignatureProperties,
+    TracingProperties,
+};
+pub use domain::cli::CliFields;
+pub use error::NotaryServerError;
+pub use server::{read_pem_file, run_tcp_server};
+pub use server_tracing::init_tracing;
+pub use util::parse_config_file;
